add: tailwind to ScrollTop
